Updated ajax, cleaned debug notices, & submit btn
So this actually takes the 3 pending pull requests from me and Kaiser and rolls them into one. I had to update how we were calling the ajax nonce and actions as submitting menu items was returning a 0. So with this, all the notices and issues should be all cleaned up and ready to be released.
To recap this includes:
* https://github.com/franz-josef-kaiser/WordPress-Post-Type-Archive-Links/commit/16aed73ef52b8019eb6d2c7a3029816f888e6ba0
* https://github.com/ryanurban/WordPress-Post-Type-Archive-Links/commit/7f35216479aa0e93500cdfd0ef6e41d35c99d655
* https://github.com/ryanurban/WordPress-Post-Type-Archive-Links/commit/7deef73a58f88a0b7dc2c7e76c4ed4f254f139c7
* And finally fixing the broken ajax actions so menu items will get appended to our menu
* Made the function open brackets and array arguement commas more WordPressy